backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
langchain-google-genai
Pillow
tensorflow
langchain
langchain-core 
langchain-community
opencv-python
python-multipart
supabase
python-dotenv 
langchain-tavily
docling